DEV - Add policies

+++ Server/app/Http/Controllers/User.php
@@ -2,8 +2,11 @@
 
 namespace App\Http\Controllers;
 
+use App\Enums\UserRoles;
 use App\Http\Resources\User as ResourcesUser;
 use App\Models\User as ModelsUser;
+use App\Policies\UserPolicy;
+use Illuminate\Auth\Access\Gate;
 use Illuminate\Http\Request;
 
 class User extends Controller
@@ -22,8 +25,14 @@ class User extends Controller
 
     public function update(Request $request, User $user) {}
 
-    public function destroy(ModelsUser $user)
+    public function destroy(Request $request, ModelsUser $user)
     {
+        if ($request->user()->cannot('delete', $user)) {
+            return response()->json([
+                'message' => 'Vous devez disposer des droit suffisant pour effectuer cette action',
+            ], 403);
+        }
+
         $user->delete();
 
         return response()->noContent();
